+++ json_structs_test.go
@@ -7,14 +7,23 @@ import (
 
 func Test_albumImage_buildFilename(t *testing.T) {
 	type fields struct {
-		FileName    string
-		ImageKey    string
-		ArchivedMD5 string
-		UploadKey   string
+		FileName         string
+		ImageKey         string
+		ArchivedMD5      string
+		UploadKey        string
+		DateTimeOriginal string
 	}
 
 	f := fields{
-		FileName:    "FileNameValue",
+		FileName:         "ti_7095_R.jpg",
+		ImageKey:         "ImageKeyValue",
+		ArchivedMD5:      "ArchivedMD5Value",
+		UploadKey:        "UploadKeyValue",
+		DateTimeOriginal: "2012-06-06T21:08:48+00:00",
+	}
+
+	f_wo_date := fields{
+		FileName:    "ti_7095_R.jpg",
 		ImageKey:    "ImageKeyValue",
 		ArchivedMD5: "ArchivedMD5Value",
 		UploadKey:   "UploadKeyValue",
@@ -31,7 +40,7 @@ func Test_albumImage_buildFilename(t *testing.T) {
 			name:         "filename",
 			fields:       f,
 			filenameConf: "{{.FileName}}",
-			want:         "FileNameValue",
+			want:         "ti_7095_R.jpg",
 			wantErr:      false,
 		},
 		{
@@ -68,14 +77,35 @@ func Test_albumImage_buildFilename(t *testing.T) {
 			name:         "complex",
 			fields:       f,
 			filenameConf: "{{.ImageKey}}-{{.FileName}}",
-			want:         "ImageKeyValue-FileNameValue",
+			want:         "ImageKeyValue-ti_7095_R.jpg",
 			wantErr:      false,
 		},
 		{
 			name:         "all",
 			fields:       f,
 			filenameConf: "prefix-{{.UploadKey}}/{{.ImageKey}}-{{.FileName}}_{{.ArchivedMD5}}",
-			want:         "prefix-UploadKeyValue/ImageKeyValue-FileNameValue_ArchivedMD5Value",
+			want:         "prefix-UploadKeyValue/ImageKeyValue-ti_7095_R.jpg_ArchivedMD5Value",
+			wantErr:      false,
+		},
+		{
+			name:         "date and time",
+			fields:       f,
+			filenameConf: "prefix-{{.Date}}/{{.Time}}",
+			want:         "prefix-2012-06-06/21_08_48",
+			wantErr:      false,
+		},
+		{
+			name:         "date and time but empty",
+			fields:       f_wo_date,
+			filenameConf: "prefix-{{.UploadKey}}/{{.ImageKey}}-{{.Date}}/{{.Time}}",
+			want:         "prefix-UploadKeyValue/ImageKeyValue-/",
+			wantErr:      false,
+		},
+		{
+			name:         "Filename with FileNameNoExt and Extension",
+			fields:       f_wo_date,
+			filenameConf: "prefix-{{.FileName}}_{{.FileNameNoExt}}_{{.Extension}}",
+			want:         "prefix-ti_7095_R.jpg_ti_7095_R_.jpg",
 			wantErr:      false,
 		},
 	}
@@ -83,11 +113,13 @@ func Test_albumImage_buildFilename(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			a := &albumImage{
-				FileName:    tt.fields.FileName,
-				ImageKey:    tt.fields.ImageKey,
-				ArchivedMD5: tt.fields.ArchivedMD5,
-				UploadKey:   tt.fields.UploadKey,
+				FileName:         tt.fields.FileName,
+				ImageKey:         tt.fields.ImageKey,
+				ArchivedMD5:      tt.fields.ArchivedMD5,
+				UploadKey:        tt.fields.UploadKey,
+				DateTimeOriginal: tt.fields.DateTimeOriginal,
 			}
+
 			tmpl, err := template.New("image_filename").Option("missingkey=error").Parse(tt.filenameConf)
 			if err != nil {
 				t.Fatal(err)