+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--
+- Add `FileNameNoExt`, `Extension`, `Date` and `Time` support in `store.file_names`
 
 ### Changed